requirements.txt
numpy
pandas
streamlit
streamlit_option_menu
streamlit_extras
PyPDF2
langchain
google-generativeai
tiktoken
faiss-cpu
selenium
